requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
greenlet==3.2.2
idna==3.10
playwright==1.52.0
pyee==13.0.0
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
